build(@mongodb-js/compass-serverstats): Refactor serverstats plugin to use new webpack config
