--- packages/compass-serverstats/.browserslistrc
@@ -1,1 +1,0 @@
-Electron >= 6
