--- packages/compass-serverstats/browser.js
@@ -1,3 +1,0 @@
-module.exports.d3 = require('./lib/d3/');
-module.exports.activate = require('./index').activate;
-module.exports.deactivate = require('./index').deactivate;
