+++ packages/compass-serverstats/electron/main/index.js
@@ -1,18 +1,20 @@
 const electron = require('electron');
+const path = require('path');
 const app = electron.app;
 const BrowserWindow = electron.BrowserWindow;
-const watch = require('watch');
-const exec = require('child_process').exec;
-const path = require('path');
-
-const ROOT_DIR = path.resolve(`${__dirname}/../..`);
-const SRC_DIR = path.join(ROOT_DIR, 'src');
 
 let mainWindow;
 
 function createWindow() {
-  mainWindow = new BrowserWindow({ width: 1200, height: 800, webPreferences: {nodeIntegration: true }});
-  mainWindow.loadURL(`file://${__dirname}/../renderer/index.html`);
+  mainWindow = new BrowserWindow({
+    width: 1200,
+    height: 800,
+    webPreferences: { nodeIntegration: true }
+  });
+  mainWindow.loadURL(
+    process.env.COMPASS_INDEX_RENDERER_URL ||
+      `file://${path.resolve(__dirname, '..', 'index.html')}`
+  );
   mainWindow.on('closed', () => {
     mainWindow = null;
   });
@@ -20,23 +22,9 @@ function createWindow() {
 
 app.on('ready', () => {
   createWindow();
-  watch.watchTree(SRC_DIR, (f, curr, prev) => {
-    if (typeof f === 'object' && prev === null && curr === null) {
-      // don't reload on initial load
-      return;
-    }
-    exec('npm run compile', {cwd: ROOT_DIR}, (err) => {
-      if (err) {
-        /* eslint no-console: 0 */
-        console.error(err);
-      }
-      mainWindow.reload();
-    });
-  });
 });
 
 app.on('window-all-closed', () => {
-  watch.unwatchTree(SRC_DIR);
   app.quit();
 });
