+++ packages/compass-serverstats/electron/renderer/index.html
@@ -1,13 +1,10 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <meta charset="UTF-8">
+    <meta charset="UTF-8" />
     <title>Component</title>
-    <link rel="stylesheet" type="text/css" href="../../styles/index.css">
   </head>
   <body>
-    <div id="container" class="rtss">
-    </div>
-    <script src="index.js" charset="UTF-8" async></script>
+    <div id="container" class="rtss"></div>
   </body>
 </html>
