+++ packages/compass-serverstats/electron/renderer/index.js
@@ -1,7 +1,7 @@
-require('@babel/register')({
-  extensions: ['.jsx'],
-  root: require('path').resolve(__dirname, '..', '..')
-});
+// Import global less file. Note: these styles WILL NOT be used in compass, as compass provides its own set
+// of global styles. If you are wishing to style a given component, you should be writing a less file per
+// component as per the CSS Modules ICSS spec. @see src/components/toggle-button for an example.
+import 'mongodb-compass/src/app/styles/index.less';
 
 const React = require('react');
 const ReactDOM = require('react-dom');
@@ -10,20 +10,20 @@ const Connection = require('mongodb-connection-model');
 const app = require('hadron-app');
 const AppRegistry = require('hadron-app-registry');
 
-const ServerStatsStore = require('../../lib/stores/server-stats-graphs-store');
-const CurrentOpStore = require('../../lib/stores/current-op-store');
-const TopStore = require('../../lib/stores/top-store');
+const ServerStatsStore = require('../../src/stores/server-stats-graphs-store');
+const CurrentOpStore = require('../../src/stores/current-op-store');
+const TopStore = require('../../src/stores/top-store');
 
 // const debug = require('debug')('mongodb-compass:server-stats');
 
 const CONNECTION = new Connection({
   hostname: '127.0.0.1',
-  port: 27017,
+  port: 27020,
   ns: 'server-stats',
   mongodb_database_name: 'admin'
 });
 
-const PerformanceComponent = require('../../lib/components');
+const PerformanceComponent = require('../../src/components');
 
 global.hadronApp = app;
 global.hadronApp.appRegistry = new AppRegistry();
