+++ packages/compass-serverstats/package.json
@@ -4,20 +4,21 @@
   "version": "15.13.0",
   "apiVersion": "3.0.0",
   "description": "Compass Real Time Server Stats Component.",
-  "main": "index.js",
-  "browser": "browser.js",
+  "main": "lib/index.js",
+  "browser": "lib/browser.js",
   "scripts": {
-    "compile": "npm run jsx && npm run less",
-    "jsx": "babel ./src --out-dir ./lib",
-    "less": "lessc ./styles/index.less ./styles/index.css",
-    "prestart": "npm run compile",
-    "start": "electron ./electron",
+    "compile": "npm run webpack -- --mode production",
+    "prewebpack": "rm -rf ./lib",
+    "webpack": "webpack-compass",
+    "prestart": "electron-rebuild --only keytar",
+    "start": "npm run webpack serve -- --mode development",
+    "poststart": "node ../../scripts/rebuild.js keytar",
     "test": "xvfb-maybe hadron-build test",
     "check": "npm run lint && npm run depcheck",
     "test-unit": "npm test -- --unit",
     "test-enzyme": "npm test -- --enzyme",
     "test-functional": "npm test -- --functional",
-    "lint": "eslint \"./src/**/*.{js,jsx}\" \"./test/**/*.js\" index.js",
+    "lint": "eslint \"./{src,test}/**/*.{js,jsx}\"",
     "depcheck": "depcheck",
     "test-ci": "npm run test",
     "prepublishOnly": "npm run compile",
@@ -25,6 +26,7 @@
   },
   "license": "SSPL",
   "peerDependencies": {
+    "d3": "^3.5.17",
     "hadron-react-buttons": "*",
     "hadron-react-components": "*",
     "react": "^16.14.0",
@@ -42,13 +44,9 @@
     "reflux": "^0.4.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.3",
     "@babel/core": "^7.14.3",
-    "@babel/plugin-proposal-decorators": "^7.14.2",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-react": "^7.13.13",
     "@babel/register": "^7.13.16",
+    "@mongodb-js/webpack-config-compass": "^0.1.0",
     "babel-loader": "^8.2.2",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
@@ -83,7 +81,6 @@
     "reflux-state-mixin": "github:mongodb-js/reflux-state-mixin",
     "sinon": "^1.17.6",
     "sinon-chai": "^2.8.0",
-    "watch": "^0.19.2",
     "xvfb-maybe": "^0.2.1"
   },
   "homepage": "https://github.com/mongodb-js/compass",
