+++ packages/compass-serverstats/webpack.config.js
@@ -1,0 +1,2 @@
+const { compassPluginConfig } = require('@mongodb-js/webpack-config-compass');
+module.exports = compassPluginConfig;
